// Desktop/loginangular/loginfron/src/app/models/product.model.ts
export interface Product {
    id: number;            // Unique identifier for the product
    name: string;          // Name of the product
    price: number;         // Price of the product
    description: string;   // Description of the product
    imageUrl: string;      // URL of the product's image
    category: string;      // Category to which the product belongs
    stock: number;         // Number of items in stock
  }
  